Fix for lcm workspace loading issue
- added check if the clientWorkspaceContext is initialized
- if clientWorkspaceContext is initialized, Dashboard component is going to wait until workspace id is resolved

JIRA: RAIL-4085
